getting the version number from git tag
