+++ .gitignore
@@ -1,0 +1,3 @@
+# Generated version file
+lib/version/VERSION
+
