+++ Makefile
@@ -1,27 +1,30 @@
-all:
+generate:
+	@git describe --tags --always --match 'v[0-9]*.[0-9]*.[0-9]*' > lib/version/VERSION
+
+all: generate
 	CGO_ENABLED=0 go install ./cmd/*
 	@cd c; make
 	go vet -composites=false ./cmd/*
 
-build-darwin:
+build-darwin: generate
 	(CGO_ENABLED=0 GOOS=darwin go build ./cmd/*)
 
-build-linux:
+build-linux: generate
 	(CGO_ENABLED=0 GOOS=linux go build ./cmd/*)
 
-build-windows:
+build-windows: generate
 	(CGO_ENABLED=0 GOOS=windows go build ./cmd/*)
 
-install-darwin:
+install-darwin: generate
 	(CGO_ENABLED=0 GOOS=darwin go install ./cmd/*)
 
-install-linux:
+install-linux: generate
 	(CGO_ENABLED=0 GOOS=linux go install ./cmd/*)
 
-install-linux-arm:
+install-linux-arm: generate
 	(CGO_ENABLED=0 GOARCH=arm64 GOOS=linux go install ./cmd/*)
 
-install-windows:
+install-windows: generate
 	(CGO_ENABLED=0 GOOS=windows go install ./cmd/*)
 
 disruption-manager.tarball:
