+++ lib/log/serverlogger/impl.go
@@ -16,6 +16,7 @@ import (
 	"github.com/Cloud-Foundations/Dominator/lib/logbuf"
 	"github.com/Cloud-Foundations/Dominator/lib/srpc"
 	"github.com/Cloud-Foundations/Dominator/lib/srpc/serverutil"
+	"github.com/Cloud-Foundations/Dominator/lib/version"
 	proto "github.com/Cloud-Foundations/Dominator/proto/logger"
 )
 
@@ -78,7 +79,9 @@ func newLogger(name string, options logbuf.Options, flags int) *Logger {
 		logger.Println("New log directory")
 	}
 	if *logAtStartup {
-		logger.Println("Startup")
+		info := version.Get()
+		logger.Printf("Startup: %s (commit: %s, built: %s, go: %s)",
+			info.Version, info.GitCommit, info.BuildDate, info.GoVersion)
 	}
 	return logger
 }
