--- lib/version/VERSION
@@ -1,1 +1,0 @@
-v0.12.0
